+++ frontend/src/components/ServicesComponents/ServicesComponentForm/ServicesComponentForm.jsx
@@ -5,7 +5,7 @@ import ServicesComponentSelect from "../ServicesComponentSelect";
 import PoPStartDate from "../PoPStartDate";
 import PoPEndDate from "../PoPEndDate";
 import TextArea from "../../UI/Form/TextArea";
-import FormHeader from "../FormHeader";
+import FormHeader from "../../UI/Form/FormHeader";
 import { NON_SEVERABLE_OPTIONS, SEVERABLE_OPTIONS, SERVICE_REQ_TYPES } from "../servicesComponents.constants";
 
 /**
